# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple quotes API with multiple categories written in GO.",
        "title": "Quotesmonger API",
        "contact": {
            "name": "adeisagopher",
            "url": "https://github.com/adeisagopher",
            "email": "adeisagopher@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/quote": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "post a quote to the database.",
                "operationId": "post-quote",
                "parameters": [
                    {
                        "description": "quote data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Quotes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Quotes"
                        }
                    }
                }
            }
        },
        "/quote/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get a quote by ID.",
                "operationId": "get-quote-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "quote ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "arrays"
                        }
                    }
                }
            }
        },
        "/quotes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "return all quotes stored in the database.",
                "operationId": "get-all-quotes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "arrays"
                        }
                    }
                }
            }
        },
        "/quotes/category": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get a quote by category.",
                "operationId": "get-quote-by-category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "quote category name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "arrays"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Quotes": {
            "type": "object"
        }
    }
}